Refactor stage.py into batch_stage.py
